Makefile: Use ccache for build
It significantly improves building times by caching and using already
built code. Even if kernel is not so big and compiles rapidly on
powerful PCs, it's better to use this flag as the constant for weak
hardware, like mine.

Tests showed 8x faster compiling on Intel Core i5-2450M comparing with
native building:

-- Native building (linux-3.4-hammerhead) -> approx. 768 secs.
-- Building with ccache (linux-3.4-hammerhead) -> approx. 96 secs.

The conclusion is obvious. The only issue is that user must have a
ccache package installed on his PC. Well, I don't think it's a big
problem, because ccache package is easy to install.

Change-Id: I4a322b0649a3a1acb33271fb07f8304f60744b9f
Signed-off-by: Alex Saiko <solcmdr@gmail.com>